- Summary: Run report in 11SP4F1 and it runs, in iHub3 it fails with java.io.UTFDataFormatException: malformed input around byte 273
- Description of Issue: 
Data has null values which I believe could be causing the issue for the
data cube.

- Description of Resolution: 
Fixed totalSum function to return Double, rather than Number object.

- TED(s) Resolved:
73643

- Regression ( Yes/No ):
NO

- Code Owner Team: 
Data Access Team

- Code Reviewers: 


- Project ID:
RA 369538

- Manual Test Description: 
Verified manually in my local environment.

- Test Automated Cases Executed(Yes/No, if “No”, then explain why):


- Special Notes: 
None

+++ data/org.eclipse.birt.data.aggregation/src/org/eclipse/birt/data/aggregation/impl/TotalSum.java
@@ -123,7 +123,7 @@ public class TotalSum extends AggrFunction
 		 */
 		public Object getSummaryValue( )
 		{
-			return sum;
+			return Double.valueOf(sum.doubleValue());
 		}
 
 	}
